Consolidated away from ImageReaderBase and added discussed options.
