+++ termsaverlib/screen/img2ascii.py
@@ -32,24 +32,33 @@ The helper class available here is:
 
     * `Img2Ascii`
 """
-from termsaverlib.screen.base.imagereader import ImageReaderBase
+#
+# Python built-in modules
+#
+import os
+import queue
+import time
+from threading import Thread
+
+#
+# Internal modules
+#
+from termsaverlib.screen.base import ScreenBase
+from termsaverlib.screen.helper.typing import TypingHelperBase
+from termsaverlib.screen.helper.position import PositionHelperBase
+from termsaverlib.screen.helper.imageconverter import ImageConverter
 from termsaverlib import constants, exception
 from termsaverlib.i18n import _
 
-import os
-
-class Img2Ascii(ImageReaderBase):
+class Img2Ascii(ScreenBase, TypingHelperBase, PositionHelperBase):
     """
-    A simple screen that will display any text (hei programmer, source code!),
+    A simple screen that will display any jpg or png image,
     on screen in a typing writer animation.
-
-    From options available in `ImageReaderBase`, this will use:
-
-        * `ImageReaderBase.cleanup_per_cycle` as True
-
-        * `ImageReaderBase.cleanup_per_file` as True
     """
+    
+    path = ''
 
+    cleanup_per_file = False
 
     def _usage_options_example(self):
         """
@@ -67,14 +76,18 @@ Options:
 
  -p, --path         Sets the location to search for text-based source files.
                     This option is mandatory.
- -d, --delay        Sets the speed at which images will shift
-                    Default is 1 second.
+ -d, --delay        Sets the speed at which the image is typed out.
+                    Default is 1 second. 
  -w, --wide         The width of each 'character' of the image.
                     Default is 2 units. (Recommended left at default)
  -c, --contrast     Displays image using a contrast based character set.
  -i, --invert       Displays the image with inverted colors.
  -s, --set          Allows the use of a custom character set.
                     Default is ' .:;+=xX$&'.
+ -z, --scale        Scales the image.
+                    Default is 1.
+ -f, --framedelay   Sets the amount of time between image shifts.
+                    Default is 5 seconds
  -h, --help         Displays this help message.
 
 Examples:
@@ -107,17 +120,10 @@ Examples:
         command-line arguments.
         """
         return _("""
-You just need to provide the path to the location from where %(app_title)s will
+You just need to provide the path to the image from where %(app_title)s will
 read and display on screen.
 
-If you do not have any code in your local machine, just get some interesting
-project from the Internet, such as Django (http://www.djangoproject.com):
-
-    If you have access to git, you may download it at:
-        git clone https://github.com/django/django.git
-
-    Or, just download the zipped source and unpack it on your local machine:
-        https://www.djangoproject.com/download/
+You may also use online images by sending a url instead of image path.
 """) % {
        'app_title': constants.App.TITLE,
     }
@@ -138,13 +144,18 @@ project from the Internet, such as Django (http://www.djangoproject.com):
               this will force the screen to be cleaned (cleared) before
               each new file is displayed
         """
-        ImageReaderBase.__init__(self,
+        ScreenBase.__init__(self,
             "img2ascii",
             _("displays images in typing animation"),
             cli_opts={
-                'opts': 'hicd:p:w:s:',
-                'long_opts': ['help', 'invert', 'path=', 'wide=', 'contrast', 'set='],
+                'opts': 'hicd:p:w:s:z:f:',
+                'long_opts': ['help', 'invert', 'path=', 'wide=', 'contrast', 'set=', 'framedelay=', 'scale='],
             })
+        self.delay = 0.002
+        self.frame_delay = 5
+        # self.path = path
+        self.cleanup_per_cycle = False
+        self.options = {'wide':2}
         self.cleanup_per_cycle = True
         self.cleanup_per_file = True
         self.options = {}
@@ -166,6 +177,16 @@ project from the Internet, such as Django (http://www.djangoproject.com):
                 self.options['contrast'] = True
             elif o in ("-s", "--set"):
                 self.options['customcharset'] = a
+            elif o in ("-f", "--framedelay"):
+                try:
+                    self.option['framedelay'] = int(a)
+                except:
+                    raise exception.InvalidOptionException("framedelay")
+            elif o in ("-z", "--scale"):
+                try:
+                    self.options['scale'] = int(a)
+                except:
+                    raise exception.InvalidOptionException("scale")
             elif o in ("-d", "--delay"):
                 try:
                     # make sure argument is a valid value (float)
@@ -175,7 +196,7 @@ project from the Internet, such as Django (http://www.djangoproject.com):
             elif o in ("-p", "--path"):
                 # make sure argument is a valid value (existing path)
                 self.path = a
-                if not os.path.exists(self.path):
+                if not os.path.exists(self.path) and self.path[0:4].lower() != 'http':
                     raise exception.PathNotFoundException(self.path,
                         _("Make sure the file or directory exists."))
             else:
@@ -186,3 +207,219 @@ project from the Internet, such as Django (http://www.djangoproject.com):
         if self.path in (None, ''):
             raise exception.InvalidOptionException("path",
                 _("It is mandatory option"), help=self._message_no_path())
+
+    def _run_cycle(self):
+        """
+        Executes a \"cycle\" of this screen.
+            * The concept of \"cycle\" is no longer accurate, and is misleading.
+              this function will not return.
+        New threaded implementation:
+            * Checks if self.path is a valid path, using `os.path.exists`
+            * Assigns a new Queue to `queue_of_valid_files`
+            * Appends a new `FileScannerThread` object to a list of threads
+            * `start()`s the `FileScannerThread`
+                * `FileScannerThread` will put paths in the queue as valid
+                   file paths are found
+            * `clear_screen()`s
+            * Gets a file from `queue_of_valid_files`, removing item from queue
+            * While nextFile (empty sequences are false)
+                * As long as there is something in the queue - that is, as long
+                  as `queue.queue.get()` is able to get an object from (the)
+                  `queue_of_valid_files`, this test evaluates True.
+                * I imagine that this behaves unpredictably given a computer
+                  with __REALLY__ slow I/O
+            * Opens `nextFile` with handle-auto-closing `with` statement and
+              `typing_print()`s it
+            * Clears screen if `self.cleanup_per_file`
+            * Puts `nextFile` ON the queue
+                * Because `queue_of_valid_files.get()` REMOVES a file path
+                  from the queue, `_run_cycle()` will never reach that path
+                  again, and eventually will exhaust the queue
+                  (failing silently, with a blank screen)
+                    * A static blank screen is the antithesis of a screensaver
+                * Therefore, `queue_of_valid_files.put(nextFile)` puts the file
+                  path at the last spot in the queue
+            * Finally, another call to `queue_of_valid_files.get()` sets up
+              the next iteration in the while loop.
+        """
+        # validate path
+        if not os.path.exists(self.path) and self.path[0:4] != 'http':
+            raise exception.PathNotFoundException(self.path)
+
+        queue_of_valid_files = queue.Queue()
+
+        threads = [self.FileScannerThread(self, queue_of_valid_files, self.path)]
+        threads[-1].daemon = True
+        threads[-1].start()
+        
+        self.clear_screen()
+        nextFile = queue_of_valid_files.get()
+        while nextFile:
+            self.get_terminal_size()
+            imgconv = ImageConverter()
+            file_data = imgconv.convert_image(nextFile, self.geometry['x'], self.geometry['y'], self.options)
+            self.typing_print(file_data)
+            time.sleep(self.frame_delay)
+            if self.cleanup_per_file:
+                self.clear_screen()
+            queue_of_valid_files.put(nextFile)
+            nextFile = queue_of_valid_files.get()
+
+    def _usage_options_example(self):
+        """
+        Describe here the options and examples of this screen.
+
+        The method `_parse_args` will be handling the parsing of the options
+        documented here.
+
+        Additionally, this is dependent on the values exposed in `cli_opts`,
+        passed to this class during its instantiation. Only values properly
+        configured there will be accepted here.
+        """
+        print (_("""
+Options:
+
+ -p, --path         Sets the location to search for text-based source files.
+                    This option is mandatory.
+ -d, --delay        Sets the speed of the displaying characters.
+                    Default is 2/1000th of a second
+ -w, --wide         The width of each 'character' of the image.
+                    Default is 2 units. (Recommended left at default)
+ -c, --contrast     Displays image using a contrast based character set.
+ -i, --invert       Displays the image with inverted colors.
+ -s, --set          Allows the use of a custom character set.
+                    Default is ' .:;+=xX$&'.
+ -z, --scale        Scales the image.
+                    Default is 1.
+ -f, --framedelay   Sets the amount of time between image shifts.
+                    Default is 5 seconds
+ -h, --help         Displays this help message.
+
+Examples:
+
+    $ %(app_name)s %(screen)s -p /path/to/images/
+    This will trigger the screensaver to read all files in the path selected
+
+    $ %(app_name)s %(screen)s -p /path/to/images/image.jpg -d 0
+    This will trigger the screensaver to read all files in the path selected
+    with no delay (too fast for a screensaver, but it's your choice that
+    matters!)
+
+    $ %(app_name)s %(screen)s -p /path/to/images/image.jpg -i -c
+    This will trigger the screensaver to read the image from the internet and
+    the image will be displayed in contrast mode and inverted.
+
+    $ %(app_name)s %(screen)s -p http://website.com/image.jpg -d 5
+    This will trigger the screensaver to read the image from the internet and
+    display it extremely slowly
+""") % {
+        'screen': self.name,
+        'app_name': constants.App.NAME,
+    })
+
+    def _recurse_to_exec(self, path, func, filetype=''):
+        """
+        Executes a function for each file found recursively within the
+        specified path.
+
+        Arguments:
+
+            * path: the path to be recursively checked (directory)
+
+            * func: the function to be executed with the file(s)
+
+            * filetype: to filter for a specific filetype
+        """
+        try:
+            if os.path.isdir(path):
+                for item in os.listdir(path):
+                    f = os.path.join(path, item)
+                    if os.path.isdir(f):
+                        if not item.startswith('.'):
+                            self._recurse_to_exec(f, func, filetype)
+                    elif f.endswith(filetype) and self._is_path_binary(f):
+                        func(f)
+            elif path.endswith(filetype) and self._is_path_binary(path):
+                func(path)
+        except:
+            # If IOError, don't put on queue, as the path might throw
+            # another IOError during screen saver operations.
+            return
+
+    @staticmethod
+    def recursively_populate_queue(self, queue_of_valid_files, path, filetype=''):
+        """
+        Populates an (empty) queue of all files within directory
+        in "path", with the paths to said files.
+
+        MUST be a staticmethod for threaded implementation to function.
+
+        Arguments:
+            * queue_of_valid_files
+
+            * path: the path to be recursively checked (directory)
+
+            * filetype: to filter for a specific filetype
+        """
+        self._recurse_to_exec(path, queue_of_valid_files.put, filetype)
+
+    def _is_path_binary(self, path):
+        """
+        Returns True if the given path corresponds to a binary, or, if for any
+        reason, the file can not be accessed or opened.
+
+        For the merit of being a binary file (i.e., termsaver will not be able
+        to handle it), it is safe enough to consider the above True, as any
+        files in this situation will be simply skipped, avoiding weird errors
+        being thrown to the end-user.
+
+        Arguments:
+
+            * path: the file location
+        """
+        CHUNKSIZE = 1024
+
+        f = None
+        try:
+            f = open(path, 'rb')
+        except:
+            # If IOError, don't even bother, as the path might throw
+            # another IOError during screen saver operations.
+            return True
+        try:
+            while True:
+                chunk = f.read(CHUNKSIZE)
+                if '\0' in chunk:  # found null byte
+                    return True
+                if len(chunk) < CHUNKSIZE:
+                    break  # done
+        except:
+            # If IOError, don't even bother, as the path might throw
+            # another IOError during screen saver operations.
+            return True
+        finally:
+            if f:
+                f.close()
+        return False
+
+    def _message_no_path(self):
+        """
+        Defines a method to be overriden by inheriting classes, with the
+        purpose to display extra help information for specific errors.
+        """
+        return ""
+
+    class FileScannerThread(Thread):
+        """Screen-animation independent thread for path scanning.
+           Allows animation to begin prior to completion of path scanning.
+        """
+        def __init__(self, fileReaderInstance, queue_of_valid_files, path_to_scan):
+            Thread.__init__(self)
+            self.__queue_of_valid_files = queue_of_valid_files
+            self.__path_to_scan         = path_to_scan
+            self.__file_reader_instance = fileReaderInstance
+        def run(self):
+            """Thread begins executing this function on
+               call to `aThreadObject.start()`.
+            """
+            Img2Ascii.recursively_populate_queue(self.__file_reader_instance, self.__queue_of_valid_files, self.__path_to_scan)
